requirements.txt
fastmcp>=2.14,<3.0
pypco
python-dotenv
starlette
uvicorn
PyJWT[crypto] 